move the json file to a json directory
hubot expects files in the `/src/` directory to be coffeescript files;
moving them a level down stops these errors.
Also, the method for loading private json definitions was failing
because fs.accessSync does nothing if a file exists; switched to
fs.access and error handling to either load the private json or fall
back to public. This requires that the private json file be delivered to
`/src/json/`